requirements.txt
Django
djangorestframework
pandas
numpy
